requirements.txt
Flask==3.0.0
google-generativeai==0.4.1
werkzeug==3.0.1